requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
google-generativeai
pydantic
streamlit
requests
python-jose[cryptography]
python-multipart
PyJWT
yfinance
pandas
